% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aggregation}
\alias{aggregation}
\title{Return values extracted from the aggregated document}
\usage{
aggregation(index, type, aggregations)
}
\arguments{
\item{index}{A string representing the index}

\item{type}{A string representing the type}

\item{aggs}{A string representing the json aggregation parameter}
}
\description{
\code{aggregation} returns values extracted from the aggregated document
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-aggregations-bucket-terms-aggregation.html}
}

