% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bulk.R
\name{bulk}
\alias{bulk}
\alias{bulk.elasticsearch}
\title{bulk}
\usage{
bulk(client, ...)

\method{bulk}{elasticsearch}(client, body, index, type, consistency = NULL,
  replication = "sync", routing = NULL, refresh = FALSE, timeout = "1m",
  raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{client}{ElasticsearchClient}

\item{body}{Json The request body}

\item{index}{String The default index for the operations}

\item{type}{String The default type for the operations}

\item{consistency}{String Explicit write consistency setting for the operation}

\item{replication}{String Explicit replication type setting}

\item{routing}{String Specific routing value}

\item{refresh}{Logical Refresh the index after performing the operation}

\item{timeout}{Explicit operation timeout}

\item{raw}{}

\item{validate.params}{}
}
\description{
Perform many index/delete operations in a single API call
}

