% Generated by roxygen2 (4.0.2): do not edit by hand
\name{count}
\alias{count}
\alias{count.elasticsearch}
\title{Count items that match a query.}
\usage{
count(client, index, type, query)

\method{count}{elasticsearch}(client, index, type, query, raw = FALSE)
}
\arguments{
\item{index}{A string representing the index.}

\item{type}{A string representing the type.}

\item{query}{A string representing the query}
}
\description{
\code{count} counts items that match a query.
}

