% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create}
\alias{create}
\alias{create.elasticsearch}
\title{create}
\usage{
create(client, ...)

\method{create}{elasticsearch}(client, index, type, id, body,
  consistency = NULL, replication = "sync", version = NULL,
  version_type = NULL, routing = NULL, parent = NULL, timestamp = NULL,
  ttl = NULL, refresh = FALSE, timeout = "1m", raw = FALSE,
  validate_params = TRUE)
}
\arguments{
\item{client}{}

\item{index}{String The name of the index}

\item{type}{String The type of the document}

\item{id}{String Document ID}

\item{body}{Json The document}

\item{version}{Number Explicit version number for concurrency control}

\item{routing}{Date Explicit operation timeout}

\item{parent}{String ID of the parent document}

\item{timestamp}{Date Explicit timestamp for the document}

\item{ttl}{Date Expiration time for the document}

\item{refresh}{Logical Refresh the index after performing the operation}

\item{timeout}{Date Explicit operation timeout}

\item{raw}{}

\item{versionType}{String Specific version type}

\item{validate.params}{}
}
\description{
Adds a typed JSON document in a specific index, making it searchable. If a document with the same index, type, and id already exists, an error will occur.
}

