% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exists.R
\name{exists}
\alias{exists}
\alias{exists.elasticsearch}
\title{exists}
\usage{
exists(client, ...)

\method{exists}{elasticsearch}(client, index, type = "_all", id,
  parent = NULL, realtime = TRUE, routing = NULL, preference = "random",
  refresh = FALSE, version = NULL, version_type = NULL, raw = FALSE,
  validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{type}{String The type of the document (use _all to fetch the first document matching the ID across all types)}

\item{id}{String The document ID}

\item{parent}{String The ID of the parent document}

\item{realtime}{Logical Specify whether to perform the operation in realtime or search mode}

\item{routing}{String Specific routing value}

\item{preference}{String Specify the node or shard the operation should be performed on (default: random)}

\item{refresh}{Logical Refresh the shard containing the document before performing the operation}

\item{version}{Number Explicit version number for concurrency control}

\item{version_type}{String Specific version type}
}
\description{
Returns a boolean indicating whether or not a given document exists.
}

