% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_source.R
\name{getSource}
\alias{getSource}
\alias{getSource.elasticsearch}
\title{getSource}
\usage{
getSource(client, ...)

\method{getSource}{elasticsearch}(client, index, type = "_all", id,
  parent = NULL, source = TRUE, realtime = TRUE, routing = NULL,
  preference = "random", refresh = FALSE, version = NULL,
  version_type = NULL, raw = FALSE, validate_params = TRUE,
  source_include = NULL, source_exclude = NULL)
}
\arguments{
\item{index}{String The name of the index}

\item{type}{String The type of the document; use _all to fetch the first document matching the ID across all types}

\item{id}{String The document ID}

\item{parent}{String The ID of the parent document}

\item{source}{String, String[], Logical True or false to return the _source field or not, or a list of fields to return}

\item{realtime}{Logical Specify whether to perform the operation in realtime or search mode}

\item{routing}{String Specific routing value}

\item{preference}{String Specify the node or shard the operation should be performed on (default: random)}

\item{refresh}{Logical Refresh the shard containing the document before performing the operation}

\item{version}{Number Explicit version number for concurrency control}

\item{source_include}{String, String[], Logical A list of fields to extract and return from the _source field}

\item{source_exclude}{String, String[], Logical A list of fields to exclude from the returned _source field}

\item{versionType}{String Specific version type}
}
\description{
Get the source of a document by its index, type and id.
}
\details{
\code{getSource} retrieves source from index
}

