% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index.elasticsearch}
\title{index}
\usage{
index(client, ...)

\method{index}{elasticsearch}(client, index, type, id, body,
  consistency = NULL, replication = "sync", version = NULL,
  version_type = NULL, op_type = NULL, routing = NULL, parent = NULL,
  timestamp = NULL, ttl = NULL, refresh = FALSE, timeout = "1m",
  raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{type}{String The type of the document}

\item{id}{String Document ID}

\item{consistency}{String Explicit write consistency setting for the operation}

\item{replication}{String Specific replication type}

\item{version}{Number Explicit version number for concurrency control}

\item{version_type}{String Specific version type}

\item{routing}{String Specific routing value}

\item{parent}{String ID of the parent document}

\item{timestamp}{String Explicit timestamp for the document}

\item{ttl}{Date Expiration time for the document}

\item{refresh}{Logical Refresh the index after performing the operation}

\item{timeout}{Date Explicit operation timeout}

\item{document}{Json}
}
\description{
Stores a typed JSON document in an index, making it searchable. When the id param is not set, a unique id will be auto-generated. When you specify an id either a new document will be created, or an existing document will be updated. To enforce "put-if-absent" behavior set the op_type to "create" or use the create() method.
}
\details{
\code{index} adds or updates a document in a specific index.
}
\examples{
index("twitter", "tweet", 1, '{"test_index": "text_index"}')
index("twitter", "tweet", body='{"test_index": "text_index"}')
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-index_.html}
}

