% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indices_analyze.R
\name{indices.analyze}
\alias{indices.analyze}
\alias{indices.analyze.elasticsearch}
\title{indices.analyze}
\usage{
indices.analyze(client, ...)

\method{indices.analyze}{elasticsearch}(client, text, index, analyzer,
  char_filters, field, filters, format, tokenizer, raw = FALSE,
  validate_params = TRUE)
}
\arguments{
\item{text}{String The text on which the analysis should be performed (when request body is not used)}

\item{index}{String The name of the index to scope the operation}

\item{analyzer}{String The name of the analyzer to use}

\item{char_filters}{String, String[], Logical A comma-separated list of character filters to use for the analysis}

\item{field}{String Use the analyzer configured for this field (instead of passing the analyzer name)}

\item{filters}{String, String[], Logical A comma-separated list of filters to use for the analysis}

\item{format}{String Format of the output (detailed or text)}

\item{tokenizer}{The name of the tokenizer to use for the analysis}

\item{raw}{Logical}

\item{validate_params}{Logical}
}
\description{
Perform the analysis process on a text and return the tokens breakdown of the text.
}
\examples{
indices.analyze(es, "will you analyse this for me please ?")
indices.analyze(es, index = "test", text = "will you analyse this for me please ?", format="detailed")
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices-analyze.html}
}

