% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indices_clear_cache.R
\name{indices.clearCache}
\alias{indices.clearCache}
\alias{indices.clearCache.elasticsearch}
\title{indices.clearCache}
\usage{
indices.clearCache(client, ...)

\method{indices.clearCache}{elasticsearch}(client, index, allow_no_indices,
  expand_wildcards, fields, field_data, filter, id_cache, ignore_unavailable,
  query_cache, raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String, String[], Logical A comma-separated list of index name to limit the operation}

\item{allow_no_indices}{Logical Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)}

\item{expand_wildcards}{String Whether to expand wildcard expression to concrete indices that are open, closed or both}

\item{fields}{String, String[], Logical A comma-separated list of fields to clear when using the field_data parameter (default: all)}

\item{field_data}{Logical Clear field data}

\item{filter}{Logical Clear filter caches}

\item{id_cache}{Logical Clear ID caches for parent/child}

\item{ignore_unavailable}{Logical Whether specified concrete indices should be ignored when unavailable (missing or closed)}

\item{query_cache}{Logical Clear query cache}

\item{raw}{Logical}

\item{validate_params}{Logical}
}
\description{
Clear either all caches or specific cached associated with one ore more indices.
}
\examples{
indices.clearCache(es)
indices.clearCache(es, "twitter")
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices-clearcache.html}
}

