% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indices.create}
\alias{indices.create}
\alias{indices.create.elasticsearch}
\title{indices.create}
\usage{
indices.create(client, ...)

\method{indices.create}{elasticsearch}(client, index, timeout = NULL,
  master_timeout = NULL, raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{timeout}{Date, Number Explicit operation timeout}

\item{master_timeout}{Date, Number Specify timeout for connection to master}
}
\description{
Create an index in Elasticsearch.
}

