% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indices_delete.R
\name{indices.delete}
\alias{indices.delete}
\alias{indices.delete.elasticsearch}
\title{indices.delete}
\usage{
indices.delete(client, ...)

\method{indices.delete}{elasticsearch}(client, index = "_all",
  timeout = NULL, master_timeout = NULL, raw = FALSE,
  validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{timeout}{Date, Number Explicit operation timeout}

\item{master_timeout}{Date, Number Specify timeout for connection to master}
}
\description{
Delete an index in Elasticsearch
}
\details{
\code{Delete_index} Deletes data in an index.
}
\examples{
index("twitter", "tweet")
index("twitter", "_all")
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/client/javascript-api/current/api-reference.html#api-indices-delete}
}

