% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indices_exists.R
\name{indices.exists}
\alias{indices.exists}
\alias{indices.exists.elasticsearch}
\title{indices.exists}
\usage{
indices.exists(client, ...)

\method{indices.exists}{elasticsearch}(client, index,
  ignore_unavailable = NULL, allow_no_indices = NULL,
  expand_wildcards = "open", local = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{ignore_unavailable}{Logical Whether specified concrete indices should be ignored when unavailable (missing or closed)}

\item{allow_no_indices}{Logical Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)}

\item{expand_wildcards}{String Whether to expand wildcard expression to concrete indices that are open, closed or both}

\item{local}{Logical Return local information, do not retrieve the state from master node (default: false)}
}
\description{
Return a boolean indicating whether given index exists.
}

