% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indices_flush.R
\name{indices.flush}
\alias{indices.flush}
\alias{indices.flush.elasticsearch}
\title{indices.flush}
\usage{
indices.flush(client, ...)

\method{indices.flush}{elasticsearch}(client, index, force, full,
  wait_if_ongoing, ignore_unavailable, allow_no_indices, expand_wildcards,
  raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String, String[], Logical A comma-separated list of index names; use _all or empty string for all indices}

\item{force}{Logical Whether a flush should be forced even if it is not necessarily needed}

\item{full}{Logical If set to true a new index writer is created and settings that have been changed related to the index writer will be refreshed.}

\item{wait_if_ongoing}{Logical If set to true the flush operation will block until the flush can be executed if another flush operation is already executing.}

\item{ignore_unavailable}{Logical Whether specified concrete indices should be ignored when unavailable (missing or closed)}

\item{allow_no_indices}{Logical Whether to ignore if a wildcard indices expression resolves into no concrete indices.}

\item{expand_wildcards}{String Whether to expand wildcard expression to concrete indices that are open, closed or both.}

\item{raw}{Logical}

\item{validate_params}{Logical}
}
\description{
Explicitly flush one or more indices.
}
\examples{
indices.flush(es, "test")
indices.flush(es, "test", force =  TRUE)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/indices-flush.html}
}

