% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/indices_get.R
\name{indices.get}
\alias{indices.get}
\alias{indices.get.elasticsearch}
\title{indices.get}
\usage{
indices.get(client, ...)

\method{indices.get}{elasticsearch}(client, index = "_all", local = FALSE,
  ignore_unavailable = FALSE, allow_no_indices = FALSE,
  expand_wildcards = "open", feature = NULL, raw = FALSE,
  validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{local}{Logical Return local information, do not retrieve the state from master node (default: false)}

\item{ignore_unavailable}{Logical Whether specified concrete indices should be ignored when unavailable (missing or closed)}

\item{allow_no_indices}{Logical Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)}

\item{expand_wildcards}{String Whether to expand wildcard expression to concrete indices that are open, closed or both}

\item{feature}{String, String[], Logical A comma-separated list of features}
}
\description{
indices.get
}

