% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indices.optimize}
\alias{indices.optimize}
\alias{indices.optimize.elasticsearch}
\title{indices.optimize}
\usage{
indices.optimize(client, ...)

\method{indices.optimize}{elasticsearch}(client, index = "_all",
  ignore_unavailable = NULL, allow_no_indices = NULL,
  expand_wildcards = "open", max_num_segments = NULL,
  only_expunge_deletes = FALSE, flush = TRUE, wait_for_merge = TRUE,
  raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{ignore_unavailable}{Logical Whether specified concrete indices should be ignored when unavailable (missing or closed)}

\item{allow_no_indices}{Logical Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)}

\item{expand_wildcards}{String Whether to expand wildcard expression to concrete indices that are open, closed or both}

\item{max_num_segments}{Number The number of segments the index should be merged into (default: dynamic)}

\item{only_expunge_deletes}{Logical Specify whether the operation should only expunge deleted documents}

\item{flush}{Logical Specify whether the index should be flushed after performing the operation (default: true)}

\item{wait_for_merge}{Logical Specify whether the request should block until the merge process is finished (default: true)}
}
\description{
Explicitly optimize one or more indices.
}

