% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info.elasticsearch}
\title{info}
\usage{
info(client, ...)

\method{info}{elasticsearch}(client, raw = FALSE)
}
\arguments{
\item{client}{ElasticsearchClient}

\item{raw}{Logical}
}
\description{
Get basic info from the current cluster.
}
\examples{
info(es)
info(es, raw = TRUE)
}

