% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scroll.R
\name{scroll}
\alias{scroll}
\alias{scroll.elasticsearch}
\title{scroll}
\usage{
scroll(client, ...)

\method{scroll}{elasticsearch}(client, scroll_id, scroll,
  validate_params = TRUE, raw = FALSE)
}
\arguments{
\item{client}{ElasticsearchClient}

\item{scroll_id}{String The scroll ID}

\item{scroll}{Duration Specify how long a consistent view of the index should be maintained for scrolled search}

\item{raw}{Logical}
}
\description{
Scroll a search request (retrieve the next set of results) after specifying the scroll parameter in a search() call.
}
\examples{
scroll(es, scroll_id = "c2Nhbjs2OzM0NDg1ODpzRlBLc0FXNlNyNm5JWUc1", scroll = "1m")
}

