% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search}
\alias{search}
\alias{search.elasticsearch}
\title{Execute a search query and retrieve results making the query}
\usage{
search(client, index, type, query, from = 0, size = 10, fields = NULL,
  source = NULL, default_operator = "OR", explain = FALSE,
  analyzer = NULL, timeout = NULL, source_include = NULL,
  search_type = "query_then_fetch", source_exclude = NULL,
  lowercase_expanded_terms = TRUE, analyze_wildcard = FALSE,
  validate = FALSE, raw = FALSE, validate.params = TRUE)

\method{search}{elasticsearch}(client, index, type, query, from = 0,
  size = 10, fields = NULL, source = NULL, default_operator = "OR",
  explain = FALSE, analyzer = NULL, timeout = NULL,
  source_include = NULL, search_type = "query_then_fetch",
  source_exclude = NULL, lowercase_expanded_terms = TRUE,
  analyze_wildcard = FALSE, validate = FALSE, raw = FALSE,
  validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index.}

\item{type}{A string representing the type.}

\item{query}{A string representing the query.}

\item{from}{A string representing the beginning of the interval we are interested in.}

\item{size}{A string representing the size.}

\item{fields}{A string representing the fields.}

\item{source}{A boolean representing our choice to return the contents of the _source field or not.}

\item{default_operator}{A string representing the default operator.}

\item{explain}{A boolean enabling explanation for each hit on how its score was computed.}

\item{analyzer}{A string that allows to use a document field property as the name of the analyzer that will be used to index the document.}

\item{timeout}{A string representing the value of the timeout .}

\item{source_include}{A string representing the source_include; include parameters that you want to display in the returned source.}

\item{search_type}{A string representing the type of the search.}

\item{source_exclude}{A string representing the source_exclude; filter parameters that you do not want to display in the returned source.}

\item{lowercase_expanded_terms}{A boolean indicating if the wildcard has lower case characters.}

\item{analyze_wildcard}{A boolean indicating, when true, that an attempt will be made to analyze wildcarded words before searching the term list for matching terms.}

\item{validate}{A boolean that allows a user to validate a potentially expensive query without executing it.}

\item{raw}{A boolean that indicates if the format of the response should be in json or not.}

\item{validate.params}{A boolean indicating the need to validate the passing parameters or not.}
}
\description{
\code{search} executes a search query and retrieves results making the query
}
\examples{
search("twitter", id="1")
search("twitter", "tweet", "1", 'c("name", "date")')
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-search.html#search-search}
}

