% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update}
\alias{update}
\alias{update.elasticsearch}
\title{Update a document with a script}
\usage{
update(client, index, type, id, body, routing = NULL, parent = NULL,
  timeout = "1m", refresh = FALSE, fields = NULL, version = NULL,
  validate.params = TRUE)

\method{update}{elasticsearch}(client, index, type, id, body, routing = NULL,
  parent = NULL, timeout = "1m", refresh = FALSE, fields = NULL,
  version = NULL, validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index}

\item{type}{A string representing the type}

\item{id}{A string representing the id}

\item{body}{A string representing the body}

\item{routing}{A string allowing to control the _routing aspect when indexing data and explicit routing control is required.}

\item{parent}{A string pointing to the parent type this child relates to.}

\item{timeout}{A string representing the value of the timeout}

\item{refresh}{A boolean that allows to explicitly refresh one or more index, making all operations performed since the last refresh.}

\item{fields}{A string representing the fields.}

\item{version}{A string representing the return of a version for each search hit.}

\item{validate.params}{A boolean indicating the need to validate the passing parameters or not.}
}
\description{
\code{update} updates a document based on a script provided.
}
\examples{
update("twitter", "tweet", 1)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-update.html}
}

