% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete}
\alias{delete}
\title{Delete a document from an index based on its id}
\usage{
delete(index, type, id)
}
\arguments{
\item{index}{A string representing the index under wi}

\item{type}{A string representing the index under wi}

\item{id}{A string representing the index under wi}
}
\description{
\code{delete} deletes a document from an index based on its id.
}
\examples{
delete("twitter", "tweet", 1)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-delete.html}
}

