% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete}
\alias{delete}
\alias{delete.elasticsearch}
\title{Delete a document from an index based on its id}
\usage{
delete(client, index, type, id, query = NULL, version = NULL,
  routing = NULL, parent = NULL, replication = "sync", refresh = FALSE,
  timeout = "1m", consistency = NULL, raw = FALSE,
  validate.params = TRUE)

\method{delete}{elasticsearch}(client, index, type, id, query = NULL,
  version = NULL, routing = NULL, parent = NULL, replication = "sync",
  refresh = FALSE, timeout = "1m", consistency = NULL, raw = FALSE,
  validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index}

\item{type}{A string representing the type}

\item{id}{A string representing the id}
}
\description{
\code{delete} deletes a document from an index based on its id.
}
\examples{
delete("twitter", "tweet", 1)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-delete.html}
}

