% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete}
\alias{delete}
\alias{delete.elasticsearch}
\title{delete}
\usage{
delete(client, ...)

\method{delete}{elasticsearch}(client, index, type, id, query = NULL,
  version = NULL, version_type = NULL, routing = NULL, parent = NULL,
  replication = "sync", refresh = FALSE, timeout = "1m",
  consistency = NULL, raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String The name of the index}

\item{type}{String The type of the document}

\item{id}{String The document ID}

\item{version}{Number Explicit version number for concurrency control}

\item{version_type}{String Specific version type}

\item{routing}{String Specific routing value}

\item{parent}{String ID of parent document}

\item{replication}{String Specific replication type}

\item{refresh}{Logical Refresh the index after performing the operation}

\item{timeout}{Date Explicit operation timeout}

\item{consistency}{String Specific write consistency setting for the operation}
}
\description{
Delete a typed JSON document from a specific index based on its id.
}
\details{
\code{delete} deletes a document from an index based on its id.
}
\examples{
delete("twitter", "tweet", 1)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-delete.html}
}

