% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Get}
\alias{Get}
\title{Get a document from an index based on its id}
\usage{
Get(index, type = "_all", id, fields = NULL, realtime = TRUE,
  routing = NULL, preference = NULL, refresh = FALSE, version = NULL,
  raw = FALSE, validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index under wi}

\item{type}{A string representing the index under wi}

\item{id}{A string representing the index under wi}

\item{raw}{A string representing the index under wi}

\item{field}{A string representing the index under wi}
}
\description{
\code{get} searches a document in an index based on its id and returns it.
}
\examples{
get("twitter", id="1")
get("twitter", "tweet", "1")
get("twitter", "tweet", "1", 'c("name", "date")')
get("twitter", "tweet", "1", 'c("name", "date")', TRUE)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-get.html}
}

