% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get}
\alias{get}
\title{Get a document from an index based on its id}
\usage{
get(index, type = "_all", id, fields = NULL, source = TRUE,
  source_include = NULL, source_exclude = NULL, realtime = TRUE,
  routing = NULL, preference = NULL, refresh = FALSE, version = NULL,
  exists = FALSE, raw = FALSE, validate.params = TRUE,
  get.source = FALSE)
}
\arguments{
\item{index}{A string representing the index.}

\item{type}{A string representing the type.}

\item{id}{A string representing the id.}

\item{fields}{A string representing the fields.}

\item{source}{A boolean representing our choice to return the contents of the _source field or not.}

\item{source_include}{A string representing the source_include; include parameters that you want to display in the returned source.}

\item{source_exclude}{A string representing the source_exclude; filter parameters that you do not want to display in the returned source.}

\item{realtime}{A boolean representing if the result should be realtime or not/}

\item{routing}{A string allowing to control the _routing aspect when indexing data and explicit routing control is required.}

\item{preference}{A string that controls a preference of which shard replicas to execute the search request on.}

\item{refresh}{A boolean that allows to explicitly refresh one or more index, making all operations performed since the last refresh available for search.}

\item{version}{A string representing the return of a version for each search hit.}

\item{exists}{A boolean that returns documents that have at least one non-null value in the original field.}

\item{raw}{A boolean that indicates if the format of the response should be in json or not.}

\item{validate.params}{A boolean indicating the need to validate the passing parameters or not.}
}
\description{
\code{get} searches a document in an index based on its id and returns it.
}
\examples{
get("twitter", id="1")
get("twitter", "tweet", "1")
get("twitter", "tweet", "1", 'c("name", "date")')
get("twitter", "tweet", "1", 'c("name", "date")', TRUE)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-get.html}
}

