% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get}
\alias{get}
\alias{get.elasticsearch}
\title{get}
\usage{
get(client, ...)

\method{get}{elasticsearch}(client, index, type = "_all", id, fields = NULL,
  source = TRUE, source_include = NULL, parent = NULL,
  source_exclude = NULL, realtime = TRUE, routing = NULL,
  preference = "random", refresh = FALSE, version = NULL,
  version_type = NULL, exists = FALSE, raw = FALSE,
  validate_params = TRUE, get_source = FALSE)
}
\arguments{
\item{client}{}

\item{index}{String The name of the index}

\item{type}{String The type of the document (use _all to fetch the first document matching the ID across all types)}

\item{id}{String The document ID}

\item{fields}{String, String[] A comma-separated list of fields to return in the response}

\item{source}{String, String[], Logical True or false to return the _source field or not, or a list of fields to return}

\item{source_include}{String, String[], Logical A list of fields to extract and return from the _source field}

\item{parent}{String The ID of the parent document}

\item{source_exclude}{String, String[], Logical A list of fields to exclude from the returned _source field}

\item{realtime}{Logical Specify whether to perform the operation in realtime or search mode}

\item{routing}{String Specific routing value}

\item{preference}{String Specify the node or shard the operation should be performed on (default: random)}

\item{refresh}{Logical Refresh the shard containing the document before performing the operation}

\item{version}{Number Explicit version number for concurrency control}

\item{version_type}{String Specific version type}
}
\description{
Get a typed JSON document from the index based on its id.
}
\details{
\code{get} searches a document in an index based on its id and returns it.
}
\examples{
get("twitter", id="1")
get("twitter", "tweet", "1")
get("twitter", "tweet", "1", 'c("name", "date")')
get("twitter", "tweet", "1", 'c("name", "date")', TRUE)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-get.html}
}

