% Generated by roxygen2 (4.0.2): do not edit by hand
\name{index}
\alias{index}
\alias{index.elasticsearch}
\title{Add or update a document in a specific index}
\usage{
index(client, index, type, id, document, version = NULL,
  version_type = NULL, op_type = NULL, routing = NULL, parent = NULL,
  timestamp = NULL, ttl = NULL, refresh = FALSE, timeout = "1m",
  raw = FALSE, validate.params = TRUE)

\method{index}{elasticsearch}(client, index, type, id, document,
  version = NULL, version_type = NULL, op_type = NULL, routing = NULL,
  parent = NULL, timestamp = NULL, ttl = NULL, refresh = FALSE,
  timeout = "1m", raw = FALSE, validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index}

\item{type}{A string representing the type}

\item{id}{A string representing the id}

\item{document}{A string representing the document}

\item{version}{A string representing the return of a version for each search hit.}

\item{version_type}{A string representing the version type.}

\item{op_type}{A string representing the operation type.}

\item{routing}{A string allowing to control the _routing aspect when indexing data and explicit routing control is required.}

\item{parent}{A string pointing to the parent type this child relates to.}

\item{timestamp}{A string indicating the value of the associated timestamp for this document.}

\item{ttl}{A string representing the time to live.}

\item{refresh}{A boolean that allows to explicitly refresh one or more index, making all operations performed since the last refresh.}

\item{timeout}{A string representing the value of the timeout}

\item{raw}{A boolean that indicates if the format of the response should be in json or not.}

\item{validate.params}{A boolean indicating the need to validate the passing parameters or not.}
}
\description{
\code{index} adds or updates a document in a specific index.
}
\examples{
index("twitter", "tweet", 1, '{"test_index": "text_index"}')
index("twitter", "tweet", document='{"test_index": "text_index"}')
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-index_.html}
}

