% Generated by roxygen2 (4.0.2): do not edit by hand
\name{indices.delete}
\alias{indices.delete}
\title{Deletes data in an index.}
\usage{
indices.delete(index = "_all", raw = FALSE)
}
\arguments{
\item{index}{A string representing the index}

\item{timeout}{A string representing the value of the timeout}
}
\description{
\code{Delete_index} Deletes data in an index.
}
\examples{
index("twitter", "tweet")
index("twitter", "_all")
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/client/javascript-api/current/api-reference.html#api-indices-delete}
}

