% Generated by roxygen2 (4.0.2): do not edit by hand
\name{search}
\alias{search}
\alias{search.elasticsearch}
\title{search}
\usage{
search(client, ...)

\method{search}{elasticsearch}(client, index, type, body,
  indices_boost = NULL, lenient = NULL, from = 0, size = 10,
  fields = NULL, source = NULL, default_operator = "OR",
  explain = FALSE, lowercase_expanded_terms = NULL, preference = "random",
  analyzer = NULL, analyze_wildcard = FALSE, df = NULL,
  ignore_unavailable = NULL, timeout = NULL, source_include = NULL,
  allow_no_indices = FALSE, expand_wildcards = "open", search_type = NULL,
  source_exclude = NULL, terminate_after = NULL, suggest_field = NULL,
  suggest_mode = "missing", suggest_size = NULL, suggest_text = NULL,
  track_scores = NULL, query_cache = NULL, validate = FALSE,
  raw = FALSE, validate_params = TRUE)
}
\arguments{
\item{index}{String A comma-separated list of index names to search; use _all or empty string to perform the operation on all indices}

\item{type}{String A comma-separated list of document types to search; leave empty to perform the operation on all types}

\item{body}{Json}

\item{indices_boost}{String, String[], Logical Comma-separated list of index boosts}

\item{lenient}{Logical Specify whether format-based query failures (such as providing text to a numeric field) should be ignored}

\item{from}{Number Starting offset (default: 0)}

\item{size}{Number Number of hits to return (default: 10)}

\item{fields}{String, String[], Logical A comma-separated list of fields to return as part of a hit}

\item{source}{String, String[], Logical True or false to return the _source field or not, or a list of fields to return}

\item{default_operator}{String The default operator for query string query (AND or OR)}

\item{explain}{Logical Specify whether to return detailed information about score computation as part of a hit}

\item{lowercase_expanded_terms}{Logical Specify whether query terms should be lowercased}

\item{preference}{String Specify the node or shard the operation should be performed on (default: random)}

\item{analyzer}{String The analyzer to use for the query string}

\item{analyze_wildcard}{Logical Specify whether wildcard and prefix queries should be analyzed (default: false)}

\item{df}{String The field to use as default where no field prefix is given in the query string}

\item{ignore_unavailable}{String Whether specified concrete indices should be ignored when unavailable (missing or closed)}

\item{timeout}{Number Explicit operation timeout}

\item{source_include}{String, String[], Logical A list of fields to extract and return from the _source field}

\item{allow_no_indices}{Logical Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)}

\item{expand_wildcards}{String Whether to expand wildcard expression to concrete indices that are open, closed or both.}

\item{search_type}{String Search operation type}

\item{source_exclude}{String, String[], Logical A list of fields to exclude from the returned _source field}

\item{terminate_after}{Number The maximum number of documents to collect for each shard, upon reaching which the query execution will terminate early.}

\item{suggest_field}{String Specify which field to use for suggestions}

\item{suggest_mode}{String Specify suggest mode}

\item{suggest_size}{Number How many suggestions to return in response}

\item{suggest_text}{String The source text for which the suggestions should be returned}

\item{track_scores}{Logical Whether to calculate and return scores even if they are not used for sorting}

\item{query_cache}{Logical Specify if query cache should be used for this request or not, defaults to index level setting}

\item{routing}{String, String[], Boolean A comma-separated list of specific routing values}

\item{version}{Logical Specify whether to return document version as part of a hit}
}
\description{
Return documents matching a query, aggregations/facets, highlighted snippets, suggestions, and more.
\code{search} executes a search query and retrieves results making the query
}
\examples{
search("twitter", id="1")
search("twitter", "tweet", "1", 'c("name", "date")')
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-search.html#search-search}
}

