% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Search}
\alias{Search}
\title{Execute a search query and retrieve results macking the query}
\usage{
Search(index, type, query, from = 0, size = 10, fields = NULL,
  source = NULL, default_operator = "OR", explain = FALSE,
  analyzer = NULL, timeout = NULL, search_type = "query_then_fetch",
  lowercase_expanded_terms = TRUE, analyze_wildcard = FALSE,
  validate = FALSE, raw = FALSE, validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index under wi}

\item{type}{A string representing the index under wi}

\item{query}{A string representing the index under wi}

\item{from}{A string representing the index under wi}

\item{size}{A string representing the index under wi}

\item{fields}{A string representing the index under wi}
}
\description{
\code{search} executes a search query and retrieves∆ results macking the query
}
\examples{
search("twitter", id="1")
search("twitter", "tweet", "1", 'c("name", "date")')
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/search-search.html#search-search}
}

