% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update}
\alias{update}
\alias{update.elasticsearch}
\title{update}
\usage{
update(client, ...)

\method{update}{elasticsearch}(client, index, type, id, body, routing = NULL,
  parent = NULL, timeout = "1m", refresh = FALSE, fields = NULL,
  version = NULL, version_type = NULL, validate_params = TRUE,
  raw = FALSE, consistency = NULL, lang = "groovy",
  replication = "sync", retry_on_conflict = NULL, timestamp = NULL,
  ttl = NULL)
}
\arguments{
\item{index}{}

\item{type}{}

\item{id}{}

\item{body}{Json}

\item{routing}{String Specific routing value}

\item{parent}{String ID of the parent document}

\item{timeout}{Number Explicit operation timeout}

\item{refresh}{Logical Refresh the index after performing the operation}

\item{fields}{String, String[], Logical A comma-separated list of fields to return in the response}

\item{version}{Number Explicit version number for concurrency control}

\item{version_type}{String Specific version type}

\item{consistency}{String Explicit write consistency setting for the operation}

\item{lang}{String The script language (default: groovy)}

\item{replication}{String Specific replication type}

\item{retry_on_conflict}{Number Specify how many times should the operation be retried when a conflict occurs (default: 0)}

\item{timestamp}{Date Explicit timestamp for the document}

\item{ttl}{Duration Expiration time for the document}
}
\description{
Update parts of a document
\code{update} updates a document based on a script provided.
}
\examples{
update("twitter", "tweet", 1)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-update.html}
}

