% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Update}
\alias{Update}
\title{Update a document with a script}
\usage{
Update(index, type, id, body, routing = NULL, parent = NULL,
  timeout = "1m", refresh = FALSE, fields = NULL, version = NULL,
  validate.params = TRUE)
}
\arguments{
\item{index}{A string representing the index under wi}

\item{type}{A string representing the index under wi}

\item{id}{A string representing the index under wi}

\item{script}{A string representing the index under wi}
}
\description{
\code{update} updates a document based on a script provided.
}
\examples{
update("twitter", "tweet", 1)
}
\references{
\url{http://www.elasticsearch.org/guide/en/elasticsearch/reference/current/docs-update.html}
}

